factor out LeaderboardContent component
